refactor(cli): extract upgrade sandboxes action
