+++ src/lib/nemoclaw-runtime-bridge.ts
@@ -3,9 +3,7 @@
 
 /* v8 ignore start -- transitional bridge until command actions are extracted from src/nemoclaw.ts. */
 
-export interface NemoClawRuntimeBridge {
-  upgradeSandboxes: (args?: string[]) => Promise<void>;
-}
+export interface NemoClawRuntimeBridge {}
 
 let runtimeFactory = (): NemoClawRuntimeBridge => {
   const runtimeModule = require("../nemoclaw") as {
